tokens, tokenswap: move checkTokenAllowance to Tokens namespace
Add docstrings to params type and function
